@@ -10,8 +10,8 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
     "lodash": "^4.17.21",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.4.1",
     "react-scripts": "5.0.1",
@@ -19,7 +19,12 @@
     "redux-devtools-extension": "^2.13.9",
     "redux-observable": "^2.0.0",
     "rxjs": "^7.5.6",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "three": "0.144.0",
+    "three-stdlib": "^2.17.0",
+    "@ar-js-org/ar.js": "^3.4.0-alpha-rc1",
+    "@react-three/drei": "^9.32.3",
+    "@react-three/fiber": "^8.8.6"
   },
   "scripts": {
     "start": "react-scripts start",
