@@ -1,9 +1,10 @@
 import React from 'react'
 import App from './App'
-import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
+import {createRoot} from 'react-dom/client'
 
-ReactDOM.render(<BrowserRouter>
-        <App />
-    </BrowserRouter>, document.getElementById('root')
-)
+const root = createRoot(document.getElementById('root'))
+
+root.render(<BrowserRouter>
+    <App />
+</BrowserRouter>)
